+++ packages/apptive_grid_core/lib/apptive_grid_model.dart
@@ -5,6 +5,9 @@ import 'package:intl/intl.dart';
 
 part 'model/data_type.dart';
 part 'model/data_entity.dart';
+part 'model/apptive_grid_uri.dart';
+
+part 'model/entity/entity_uri.dart';
 
 part 'model/form/component/form_component.dart';
 part 'model/form/component/boolean_form_component.dart';