+++ packages/apptive_grid_core/lib/model/apptive_grid_uri.dart
@@ -1,0 +1,7 @@
+part of apptive_grid_model;
+
+/// Base class used to represent Objects with a Uri
+abstract class ApptiveGridUri {
+  /// Generates the uriString used for ApiCalls referencing this
+  String get uriString;
+}